refactor visit page
