+++ hackathon/pages/visit_page/area_selector.py
@@ -1,26 +1,26 @@
 import reflex as rx
-from hackathon.state import State
+from hackathon.pages.visit_page.state import VisitState
 
 def next_area_button() -> rx.Component:
     """The next area button."""
     return rx.button(
         "Next",
-        on_click=lambda: State.next_area,
+        on_click=lambda: VisitState.next_area,
     )
 
 def previous_area_button() -> rx.Component:
     """The previous area button."""
     return rx.button(
         "Previous",
-        on_click=lambda: State.previous_area,
+        on_click=lambda: VisitState.previous_area,
     )
 
 def select_area_dropdown() -> rx.Component:
     """The select area dropdown."""
     return rx.select(
-        items=State.areas_names,
-        value=State.current_area_name,
-        on_change=State.switch_area,
+        items=VisitState.areas_names,
+        value=VisitState.current_area_name,
+        on_change=VisitState.switch_area,
     )
 
 def area_selector_panel() -> rx.Component:
