+++ hackathon/pages/visit_page/media.py
@@ -1,13 +1,13 @@
 import reflex as rx
 
-from hackathon.state import State
+from hackathon.pages.visit_page.state import QA, VisitState
 
 
 def video() -> rx.Component:
     """A video player."""
     return (
         rx.video(
-            url=State.video_url,
+            url=VisitState.video_url,
             width="100%",
             height="100%",
         ),
